account-service setup done

+++ account-service/src/main/java/com/bingbank/accountservice/repository/BranchRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface BranchRepository extends JpaRepository<Branch, Long> {
+public interface BranchRepository extends JpaRepository<Branch, Integer> {
     
     Optional<Branch> findByBranchCode(String branchCode);
 }
